[linux] Update target headless backends
Enable OSMesa backend via WITH_OSMESA=1 envvar.

OSMesa backend uses Gallium OSMesa implementation, which uses
LLVMpipe as default driver.

+++ scripts/travis_setup.sh
@@ -30,16 +30,6 @@ ${CXX} --version
 # Ensure mason is on the PATH
 export PATH="`pwd`/.mason:${PATH}" MASON_DIR="`pwd`/.mason"
 
-# Start the mock X server
-if [ -f /etc/init.d/xvfb ] ; then
-    mapbox_time "start_xvfb" \
-    sh -e /etc/init.d/xvfb start
-    sleep 2 # sometimes, xvfb takes some time to start up
-fi
-
-# Make sure we're connecting to xvfb
-export DISPLAY=:99.0
-
 mapbox_time "checkout_mason" \
 git submodule update --init .mason
 
@@ -47,10 +37,20 @@ git submodule update --init .mason
 mapbox_time "touch_package_json" \
 touch package.json
 
-# Install and set up to load a more recent version of mesa
-mapbox_time "install_mesa" \
-mason install mesa 11.2.2
-export LD_LIBRARY_PATH="`mason prefix mesa 11.2.2`/lib:${LD_LIBRARY_PATH:-}"
+# Start the mock X server
+if [ -f /etc/init.d/xvfb ] && [ -z "${WITH_OSMESA}" ]; then
+    mapbox_time "start_xvfb" \
+    sh -e /etc/init.d/xvfb start
+    sleep 2 # sometimes, xvfb takes some time to start up
+
+    # Make sure we're connecting to xvfb
+    export DISPLAY=:99.0
+
+    # Install and set up to load a more recent version of mesa
+    mapbox_time "install_mesa" \
+    mason install mesa 11.2.2
+    export LD_LIBRARY_PATH="`mason prefix mesa 11.2.2`/lib:${LD_LIBRARY_PATH:-}"
+fi
 
 # Install and set up to load awscli
 pip install --user awscli
